+++ bot/cogs/games.py
@@ -4,15 +4,25 @@ from discord import app_commands
 import random
 import asyncio
 import json
-import os
+import logging
 from itertools import cycle
+from pathlib import Path
 from abc import ABC, abstractmethod
 from typing import Dict, List, Any
 
 from bot.utils.anime_api import fetch_random_character, build_character_select_options
-from bot.utils.game_texts import random_win_message, random_lose_message, compute_rewards, award_rewards
+from bot.utils.game_texts import (
+    random_win_message,
+    random_lose_message,
+    compute_rewards,
+    award_rewards,
+)
 from bot.utils.discord_helpers import create_same_choices
 from bot.config.emojis import CustomEmojis
+from bot.config.paths import DATA_PATH
+
+logger = logging.getLogger(__name__)
+
 
 """
 games.py
@@ -42,11 +52,13 @@ Design notes (important):
 
 # --- Trivia Data Adapter Interface ---
 
+
 class TriviaLoader(ABC):
     """
     Abstract interface for loading trivia questions.
     Allows swapping the underlying data source (JSON, DB, API) without changing game logic.
     """
+
     @abstractmethod
     async def load_data(self) -> Dict[str, List[Dict[str, Any]]]:
         """
@@ -55,13 +67,14 @@ class TriviaLoader(ABC):
         """
         pass
 
+
 class JsonTriviaLoader(TriviaLoader):
     """
     Concrete implementation of TriviaLoader for local JSON files.
     """
-    def __init__(self, file_path: str, loop: asyncio.AbstractEventLoop):
-        self.file_path = file_path
-        self.loop = loop
+
+    def __init__(self, file_path: str | Path):
+        self.file_path = Path(file_path)
 
     async def load_data(self) -> Dict[str, List[Dict[str, Any]]]:
         def _read_file():
@@ -69,14 +82,14 @@ class JsonTriviaLoader(TriviaLoader):
                 with open(self.file_path, "r", encoding="utf-8") as f:
                     return json.load(f)
             except FileNotFoundError:
-                print(f"[Games] Trivia file not found at {self.file_path}")
+                logger.warning("Trivia file not found: %s", self.file_path)
                 return {}
             except json.JSONDecodeError:
-                print(f"[Games] Invalid JSON in {self.file_path}")
+                logger.error("Invalid JSON in trivia file: %s", self.file_path)
                 return {}
 
         # Offload blocking I/O to executor
-        data = await self.loop.run_in_executor(None, _read_file)
+        data = await asyncio.to_thread(_read_file)
 
         # Normalize structure to Dict[str, List]
         if isinstance(data, dict):
@@ -86,7 +99,9 @@ class JsonTriviaLoader(TriviaLoader):
         else:
             return {"Mixed": []}
 
+
 # --- Games Cog ---
+
 
 class Games(commands.Cog):
     """
@@ -97,16 +112,17 @@ class Games(commands.Cog):
     - Present per-question UI, handle user answers, and reward correct responses.
     - Integrate with anime_api to provide image-based character guessing.
     """
+
     def __init__(self, bot):
         self.bot = bot
-        
-        # Determine path and initialize the loader adapter
-        data_path = os.path.join(os.path.dirname(__file__), "..", "data", "trivia.json")
-        self.trivia_loader: TriviaLoader = JsonTriviaLoader(data_path, bot.loop)
-        
+
+        # Use the canonical repository-level data directory.
+        trivia_path = DATA_PATH / "trivia.json"
+        self.trivia_loader: TriviaLoader = JsonTriviaLoader(trivia_path)
+
         # Initialize with safe defaults; data is loaded in cog_load
         self.trivia_dict = {"Mixed": []}
-        
+
         # used_questions prevents immediate repetition across quiz runs; it's intentionally
         # kept in-memory so it resets on bot restart.
         self.used_questions = set()
@@ -116,9 +132,9 @@ class Games(commands.Cog):
         Asynchronously load trivia data using the configured loader.
         """
         self.trivia_dict = await self.trivia_loader.load_data()
-            
+
         total_q = sum(len(qs) for qs in self.trivia_dict.values())
-        print(f"[Games] Loaded {total_q} trivia questions.")
+        logger.info("Loaded %d trivia questions.", total_q)
 
     def get_balanced_questions(self, num_questions: int):
         """
@@ -145,7 +161,11 @@ class Games(commands.Cog):
 
         while len(questions) < num_questions:
             title = next(title_cycle)
-            available = [q for q in self.trivia_dict[title] if q["question"] not in self.used_questions]
+            available = [
+                q
+                for q in self.trivia_dict[title]
+                if q["question"] not in self.used_questions
+            ]
             if available:
                 q = random.choice(available)
                 self.used_questions.add(q["question"])
@@ -165,7 +185,7 @@ class Games(commands.Cog):
         *,
         exp_mul=(2, 3),
         exp_base=(5, 10),
-        coin_range=(15, 30)
+        coin_range=(15, 30),
     ):
         """
         Handle awarding rewards for a correct answer.
@@ -193,7 +213,9 @@ class Games(commands.Cog):
         await award_rewards(profile_cog, user_id, guild_id, exp_reward, coin_reward)
         await send_fn(random_win_message(exp_reward, coin_reward))
 
-    @commands.hybrid_command(name="animequiz", description="Start an anime trivia quiz.")
+    @commands.hybrid_command(
+        name="animequiz", description="Start an anime trivia quiz."
+    )
     @commands.guild_only()
     @app_commands.describe(questions="Number of questions")
     @app_commands.choices(questions=create_same_choices([5, 10, 15, 20]))
@@ -206,7 +228,7 @@ class Games(commands.Cog):
         """
         num_questions = questions.value
         quiz_questions = self.get_balanced_questions(num_questions)
-        
+
         if not quiz_questions:
             return await ctx.send("❌ No trivia questions available.")
 
@@ -215,16 +237,20 @@ class Games(commands.Cog):
         for idx, question in enumerate(quiz_questions, 1):
             options_list = list(question["options"])
             random.shuffle(options_list)
-            options = [discord.SelectOption(label=opt, value=opt) for opt in options_list]
+            options = [
+                discord.SelectOption(label=opt, value=opt) for opt in options_list
+            ]
 
             embed = discord.Embed(
                 title=f"Question {idx}/{num_questions}",
-                description=question["question"]
+                description=question["question"],
             )
             view = discord.ui.View()
             future = asyncio.get_event_loop().create_future()
 
-            select = discord.ui.Select(placeholder="Choose an answer...", options=options)
+            select = discord.ui.Select(
+                placeholder="Choose an answer...", options=options
+            )
             view.add_item(select)
 
             async def callback(
@@ -235,7 +261,9 @@ class Games(commands.Cog):
             ):
                 # Only allow the quiz invoker to answer this question.
                 if interaction.user != ctx.author:
-                    await interaction.response.send_message("This is not your game!", ephemeral=True)
+                    await interaction.response.send_message(
+                        "This is not your game!", ephemeral=True
+                    )
                     return
                 if not _future.done():
                     _future.set_result(interaction.data["values"][0])
@@ -251,8 +279,10 @@ class Games(commands.Cog):
 
                 if selected == question["answer"]:
                     score += 1
+
                     async def send_ctx(msg: str):
                         await ctx.send(msg)
+
                     await self._handle_correct_answer(
                         ctx.author.id,
                         ctx.guild.id,
@@ -267,28 +297,36 @@ class Games(commands.Cog):
             except asyncio.TimeoutError:
                 select.disabled = True
                 await message.edit(view=view)
-                await ctx.send(f"{CustomEmojis['TIME']} Time's up! The correct answer was `{question['answer']}`.")
+                await ctx.send(
+                    f"{CustomEmojis['TIME']} Time's up! The correct answer was `{question['answer']}`."
+                )
 
         await ctx.send(f"🏁 Quiz finished! You scored **{score}/{num_questions}**.")
 
-    @commands.hybrid_command(name="guesscharacter", description="Guess a random popular anime character")
+    @commands.hybrid_command(
+        name="guesscharacter", description="Guess a random popular anime character"
+    )
     @commands.guild_only()
     async def guesscharacter(self, ctx):
         """
         Image-based character guess game.
 
-        Fetches a random character from AniList/Jikan and presents a 
+        Fetches a random character from AniList/Jikan and presents a
         multiple-choice menu with the correct name and randomized distractors.
         """
         try:
             # Need a session to pass to the API; get from bot
             if not self.bot.session:
-                 return await ctx.send("❌ Bot network session not initialized.")
-            
-            character = await fetch_random_character(session=self.bot.session, prefer="AniList")
+                return await ctx.send("❌ Bot network session not initialized.")
+
+            character = await fetch_random_character(
+                session=self.bot.session, prefer="AniList"
+            )
         except Exception as e:
             print(f"[Games] guesscharacter error: {e}")
-            return await ctx.send("❌ Couldn't fetch characters from any API. Please try again later.")
+            return await ctx.send(
+                "❌ Couldn't fetch characters from any API. Please try again later."
+            )
 
         correct_name = character["name"]
         image = character["image"]
@@ -296,11 +334,17 @@ class Games(commands.Cog):
         source = character["source"]
 
         try:
-            options_list = await build_character_select_options(correct_name, source, session=self.bot.session)
+            options_list = await build_character_select_options(
+                correct_name, source, session=self.bot.session
+            )
         except Exception:
-            return await ctx.send("❌ Failed to fetch options for the quiz. Please try again.")
+            return await ctx.send(
+                "❌ Failed to fetch options for the quiz. Please try again."
+            )
 
-        embed = discord.Embed(title="Guess the character!", description=f"From **{anime_title}**")
+        embed = discord.Embed(
+            title="Guess the character!", description=f"From **{anime_title}**"
+        )
         embed.set_image(url=image)
         embed.set_footer(text=f"Source: {source}")
 
@@ -309,7 +353,9 @@ class Games(commands.Cog):
         view.anime_title = anime_title
         view.author_id = ctx.author.id
 
-        select = discord.ui.Select(placeholder="Choose the correct character...", options=options_list)
+        select = discord.ui.Select(
+            placeholder="Choose the correct character...", options=options_list
+        )
         view.add_item(select)
 
         message = await ctx.send(embed=embed, view=view)
@@ -325,17 +371,22 @@ class Games(commands.Cog):
         - Awards rewards on correct answers via _handle_correct_answer (uses an interaction-aware sender).
         - Disables the Select and updates the message view after an answer.
         """
+
         async def callback(interaction: discord.Interaction):
             if interaction.user.id != view.author_id:
-                await interaction.response.send_message("This is not your game!", ephemeral=True)
+                await interaction.response.send_message(
+                    "This is not your game!", ephemeral=True
+                )
                 return
             selected = interaction.data["values"][0]
             correct_name = view.correct_answer
             anime_title = view.anime_title
 
             if selected == correct_name:
+
                 async def send_interaction(msg: str):
                     await interaction.response.send_message(msg)
+
                 await self._handle_correct_answer(
                     interaction.user.id,
                     interaction.guild.id,
@@ -345,13 +396,17 @@ class Games(commands.Cog):
                     coin_range=(15, 30),
                 )
             else:
-                await interaction.response.send_message(random_lose_message(correct_name, anime_title))
+                await interaction.response.send_message(
+                    random_lose_message(correct_name, anime_title)
+                )
 
             for item in view.children:
                 if isinstance(item, discord.ui.Select):
                     item.disabled = True
             await message.edit(view=view)
+
         return callback
+
 
 async def setup(bot):
     await bot.add_cog(Games(bot))
